+++ integration/tests/e2e_tests.rs
@@ -1,12 +1,14 @@
-use integration::test_util::{load_batch, load_chunk, load_chunk_for_test, ASSETS_DIR, PARAMS_DIR};
+use integration::{
+    prove::get_blob_from_chunks,
+    test_util::{load_batch, load_chunk, load_chunk_for_test, ASSETS_DIR, PARAMS_DIR},
+};
 use prover::{
     eth_types::H256,
     proof::dump_as_json,
     utils::{chunk_trace_to_witness_block, init_env_and_log, read_env_var},
-    zkevm, BatchData, BatchHash, BatchHeader, BatchProvingTask, ChunkInfo, ChunkProvingTask,
-    MAX_AGG_SNARKS,
+    zkevm, BatchHash, BatchHeader, BatchProvingTask, ChunkInfo, ChunkProvingTask, MAX_AGG_SNARKS,
 };
-use std::{env, fs, path::Path};
+use std::env;
 
 fn load_test_batch() -> anyhow::Result<Vec<String>> {
     let batch_dir = read_env_var("TRACE_PATH", "./tests/extra_traces/batch_25".to_string());
@@ -52,6 +54,7 @@ fn test_e2e_prove_verify() {
 
         let batch_proof = prove_and_verify_batch(&output_dir, batch_prover, batch);
         /*
+        use std::{fs, path::Path};
         let proof_path = Path::new(&output_dir).join("full_proof_batch_agg.json");
         let proof_path_to =
             Path::new(&output_dir).join(format!("full_proof_batch_agg_{}.json", i + 1).as_str());
@@ -140,27 +143,6 @@ fn test_batch_bundle_verify() -> anyhow::Result<()> {
     log::info!("bundle 4 batches OK");
 
     Ok(())
-}
-
-// `chunks` are unpadded
-// TODO: move it elsewhere?
-// Similar codes with aggregator/src/tests/aggregation.rs
-// Refactor?
-fn get_blob_from_chunks(chunks: &[ChunkInfo]) -> Vec<u8> {
-    let num_chunks = chunks.len();
-
-    let padded_chunk =
-        ChunkInfo::mock_padded_chunk_info_for_testing(chunks.last().as_ref().unwrap());
-    let chunks_with_padding = [
-        chunks.to_vec(),
-        vec![padded_chunk; MAX_AGG_SNARKS - num_chunks],
-    ]
-    .concat();
-    let batch_data = BatchData::<{ MAX_AGG_SNARKS }>::new(chunks.len(), &chunks_with_padding);
-    let batch_bytes = batch_data.get_batch_data_bytes();
-    let blob_bytes = prover::aggregator::eip4844::get_blob_bytes(&batch_bytes);
-    log::info!("blob_bytes len {}", blob_bytes.len());
-    blob_bytes
 }
 
 fn gen_batch_proving_task(
